♻️ refactor(server): extract work registration into its own service module
Move the completion-time Work registration code scattered across
services/agentRuntime — the file/shell scan orchestrator, the shell scan
engine + per-CLI scanners, and the desktop-local run scan — into a
dedicated services/workRegistration module, matching the one-folder-per-
service layout used across apps/server/src/services.

- workRegistration.ts → workRegistration/registerWorksForOperation.ts,
  with the pure stateHasEntityFileEdits predicate split into its own file
- shared UNEXECUTED_INTERVENTION_STATUSES deduplicated into constants.ts
- public surface re-exported from workRegistration/index.ts; callers now
  import @/server/services/workRegistration
- agent-work skill file map updated to the new paths
