+++ apps/server/src/services/agentRuntime/AgentRuntimeService.ts
@@ -61,6 +61,7 @@ import { QueueService } from '@/server/services/queue';
 import { LocalQueueServiceImpl } from '@/server/services/queue/impls';
 import { ToolExecutionService } from '@/server/services/toolExecution';
 import { BuiltinToolsExecutor } from '@/server/services/toolExecution/builtin';
+import { stateHasEntityFileEdits } from '@/server/services/workRegistration';
 
 import { isAbortError, throwIfAborted } from './abort';
 import {
@@ -93,7 +94,6 @@ import {
   type StepCompletionReason,
   type SubAgentBridgeParams,
 } from './types';
-import { stateHasEntityFileEdits } from './workRegistration';
 
 if (process.env.VERCEL) {
   // Route debug output to stdout (`console.info`) instead of stderr, which
