+++ apps/server/src/services/workRegistration/shellWorkRegistration.ts
